refactor: move the `UnsuccessfulFetch` class
