+++ index.ts
@@ -1,3 +1,2 @@
 export * from "./src/client.js";
-export * from "./src/error.js";
 export { default } from "./src/client.js";
