+++ src/client.ts
@@ -1,5 +1,4 @@
 import fetch, { RequestInit, Response, Headers } from "node-fetch";
-import { UnsuccessfulFetch } from "./error.js";
 
 export type ITransformType =
   | "buffer"
@@ -21,6 +20,20 @@ export interface IClientOptions extends FetchClientOptions {
 }
 
 export const DefaultTransform = "raw";
+
+export class UnsuccessfulFetch extends Error {
+  readonly #response: Response;
+
+  public constructor(message: string, response: Response) {
+    super(message);
+    this.name = "UnsuccessfulFetch";
+    this.#response = response;
+  }
+
+  public get response(): Response {
+    return this.#response;
+  }
+}
 
 export class FetchClient<T = Response> {
   #fetchOptions: RequestInit;
